+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/RealMiniAppSpec.kt
@@ -79,8 +79,10 @@ class RealMiniAppSpec {
 
             verify(miniAppDownloader, times(1)).getMiniApp(TEST_MA_ID)
             verify(displayer, times(1))
-                .createMiniAppDisplay(getMiniAppResult.first, getMiniAppResult.second,
-                    miniAppMessageBridge, null, miniAppCustomPermissionCache)
+                .createMiniAppDisplay(
+                    getMiniAppResult.first, getMiniAppResult.second,
+                    miniAppMessageBridge, null, miniAppCustomPermissionCache, ""
+                )
         }
 
     @Test
@@ -92,8 +94,14 @@ class RealMiniAppSpec {
 
             verify(miniAppDownloader, times(1)).getMiniApp(TEST_MA)
             verify(displayer, times(1))
-                .createMiniAppDisplay(getMiniAppResult.first, getMiniAppResult.second,
-                    miniAppMessageBridge, miniAppNavigator, miniAppCustomPermissionCache)
+                .createMiniAppDisplay(
+                    getMiniAppResult.first,
+                    getMiniAppResult.second,
+                    miniAppMessageBridge,
+                    miniAppNavigator,
+                    miniAppCustomPermissionCache,
+                    ""
+                )
         }
 
     @Test